16570: resource_aviatrix_firenet: Ignore egress errors
If a firenet is being used as an egress firenet then the
backend places limitations on changing egress_enabled
and inspection_enabled. The backend will return the error
'[AVXERR-FIRENET-0011] Unsupported for Egress Transit.'
if we try to change these attributes for an egress firenet.
This PR ignores this error when deleting an egress firenet.
This PR also ignores the error when the firenet is initially
created.