tweak: fix 'enable_bgp_over_lan' import
When importing a non-Azure transit gateway we would
not Set anything in state for 'enable_bgp_over_lan'
so it would be effectively set to null. However,
since it has a default value of false, it would be
read as a Diff on the next plan.

So, instead for non-Azure transit gateways we will always
set 'enable_bgp_over_lan' to it's default value of false.