19914: Fix firewall_instance.key_name import
Due to a bug in export we would not export
the key_name field. Then if a user imported
the resource while the key_name field was
blank then the key_name was not read into state.

From then on the key_name would never be read into
state, even if user added the key_name to the config
because of the blank value in the state file.