20432: Update firenet doc
Update the firenet documentation to inform users
that an explicit depends_on needs to be set on any
of their firewall_instance_association resources.

Also added an example of the depends_on to the example
since users often copy/paste directly from the example.

Note that I specifically did not include examples of
setting depends on for transit_gateway or firewall_instance
resources because the firewall_instance_association should
cover the dependency on those resources.