AVX-14274: Fix fail_close_enabled being set to incorrect value
Normally fail_close_enabled is defaulted to false, however,
in some cases (not sure all the cases) it is set to true.
To be safe, I will call the disable API if user sets
fail_close_enabled = false in the Create function. This is okay
because even if it already disabled, API will return 'configuration
not changed' error that we already ignore in the API function.