AVX-52626: Fix S2C so that it correctly updates remote_subnet_cidrs (#1985)
* AVX-52626: Fix S2C so that it correctly updates remote_subnet_cidrs

There are a few things going on here:

1) At some point it looks like we made changes so that `remote_subnet_cidr` and
   `local_subnet_cidr` were copied into `cloud_subnet_cidr` when editing an sc2
   connection. We may have, at some point, been using the network type field
   to differentiate the two. Network type isn't used as a differentiator
   so this breaks when making changes to remote or local subnet cidr.

2) Additionally it seems that for edit, we use `remote_cidr`, not
   `remote_subnet_cidr` as we do for create.

3) Remove references to NetworkType.
    This doesn't seem to be used on the controller side.

This fix ensure that only `local_subnet_cidr` uses `cloud_subnet_cidr` and
`remote_subnet_cidr` uses `remote_cidr` when updating.

Additionally this fixes other mismatched fields that are different in
when performing and edit.

Ideally we fix the API in the controller as this is a bit confusing on
mutiple levels.  We should normalize the API for create and edit and
most likely deprecate cloud_subnet_cidr or do away with it completely.

(cherry picked from commit 50d1dd319a8853151f7f5eef41dff585ec90a8dd)