AVX-49421: Cache the results of list accounts instead of making n calls
Currently we call list_accounts for every account, iterate through those results
to find a specific account and return that reuslt. This can result in hundreds
of calls if the customer has hundreds of accounts.

Instead, cache the results of the list_accounts call and use that to populate
all accounts.

Tested with ~600 accounts and this brings us from ~35 minutes to about 30 to 15
seconds.