AVX-55228: Use semver comparison from 8.0 onwards when checking versions
Currently we check version compatability between the Controller and the
Terraform version. Previous to 8.0 the version numbers were not semver
compatabile, so we had a lot of custom parsing.

We now simply check the supported version map with the current version using
semver to compare the major/minor only.

Removes ParseVersion from GetCurrentVersion and instead returns the raw version
string.

Removes unsupported gateway and controller upgrades since this won't be supported
in 8.0.

Remove dead version parsing code and tests.