AVX-61120: Fix ExternalDeviceConnBgpBfdDiffSuppressFunc. (#2149)
* AVX-61120: Fix ExternalDeviceConnBgpBfdDiffSuppressFunc.

The old implementation of ExternalDeviceConnBgpBfdDiffSuppressFunc could
suppress a diff if any of the fields were set to the default value. This
broke various cases, including trying to reconcile drift when the
desired value matched the default value.

Fixing this was way more complicated than expected, because
`GetDiff("bgp_bfd")` did not work as expected (see code comment). This
PR changes the diff suppression to rely on terraform to populate list defaults.
The net result is that we now only suppress if there is no change between
the user config and the state, while accounting for defaults.

* golint

(cherry picked from commit 4e231e762bd68061c08021eec5daa69e6d46778d)