AVX-61542: Remove GetEdgeExternalDeviceConnDetail (#2164)
AVX-61542 was raised for bgp_bfd not being "read back" correctly for edge
spokes. This is because BFD code was added to `GetExternalDeviceConnDetail`,
but _not_ GetEdgeExternalDeviceConnDetail. It turns out that the two
functions are _identical_ except for the missing BFD support, plus code
to switch to BGP parameters when `WanUnderlay` is set.

This PR removes the edge-specific API function and instead moves the
handling of the backup parameters into the edge resource Read, which is
probably the right place for that anyway.