AVX-59186: terraform changes for route approval
Prior to Release 8.1, users were only allowed to set up "exact prefix match"
values for the list of approved_learned_cidrs.  In 8.1, we have expanded that to
take optional "ge" and "le" values in order to let users specify a range of
prefixes.  We also refactored the APIs to remove unnecessary duplication, for
example, we used to have separate APIs for spoke/ transit that were doing the
same thing.  In this PR we modify the Aviatrix Terraform Provider to use these
these new APIs.