AVX-68208: Create a new resource user group membership (#2319)
The **aviatrix_rbac_group_user_membership** resource allows the
creation and management of user membership for Aviatrix  RBAC groups.
This resource is authoritative for the group's user membership,
meaning it manages the complete set of users that belong to a specific
group.

**Note:** There is another related resource,
`aviatrix_rbac_group_user_attachment` which manages a single
user-to-group attachment per resource. While that resource is still
available it would be preferable to use this new resource especially
for larger sets of users or groups.